Clean up text_stats imports and tests
