+++ src/textacy/text_stats/__init__.py
@@ -1,1 +1,3 @@
-from .api import TextStats
+from .api import TextStats, load_hyphenator
+from . import basics
+from . import readability
